api/requirements.txt
fastapi
uvicorn
requests
pillow==9.5.0
openai==0.27.6
python-dotenv
ultralytics
pytorch_lightning==1.5.10
lmdb==1.4.1
timm==0.6.13
nltk
youtube-transcript-api
pytube
tiktoken 
faiss-cpu
langchain==0.0.162
pygpt4all==1.1.0
chromadb==0.3.22
llama-cpp-python==0.1.47
pypdf
google-api-python-client
bs4
